src/main/webapp/app/entities/fast-food/route/fast-food-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IFastFood, FastFood } from '../fast-food.model';
import { FastFoodService } from '../service/fast-food.service';

import { FastFoodRoutingResolveService } from './fast-food-routing-resolve.service';

describe('FastFood routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: FastFoodRoutingResolveService;
  let service: FastFoodService;
  let resultFastFood: IFastFood | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(FastFoodRoutingResolveService);
    service = TestBed.inject(FastFoodService);
    resultFastFood = undefined;
  });

  describe('resolve', () => {
    it('should return IFastFood returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFastFood = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultFastFood).toEqual({ id: 123 });
    });

    it('should return new IFastFood if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFastFood = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultFastFood).toEqual(new FastFood());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as FastFood })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFastFood = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultFastFood).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
